Move category to recipe app.
