+++ src/ingredients/models.py
@@ -1,15 +1,8 @@
 from django.db import models
 
-class Category(models.Model):
-    name = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
-
 class Ingredient(models.Model):
     name = models.CharField(max_length=100)
     kcal_per_100g = models.IntegerField(default=0)
-    category = models.ForeignKey(Category, on_delete=models.CASCADE)
 
     def __str__(self):
         return self.name
