+++ src/ingredients/urls.py
@@ -6,5 +6,4 @@ urlpatterns =[
     path("", views.List.as_view(), name="list"),
     path("add/", views.Add.as_view(), name="add"),
     path("addToShoppingList/", views.addToShoppingList, name="addToShoppingList"),
-    path("createCategory/", views.CreateCategory.as_view(), name="createCategory"),
 ]
