+++ src/recipies/models.py
@@ -1,10 +1,17 @@
 from django.db import models
 from ingredients.models import Ingredient
 
+class Category(models.Model):
+    name = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.name
+
 class Recipe(models.Model):
     title = models.CharField(max_length=200)
     ingredients = models.ManyToManyField(Ingredient)
     instruction = models.TextField(max_length=500)
+    category = models.ForeignKey(Category, on_delete=models.CASCADE)
 
     def __str__(self):
         return self.title
