+++ src/recipies/templates/category_form.html
@@ -1,5 +1,5 @@
 <form method="POST">
     {% csrf_token %}
-        {{ form.as_p }}
+    {{ form.as_p }}
     <button type="submit">Add category</button>
-    </form>
+</form>
