+++ src/recipies/templates/recipies/recipe_form.html
@@ -1,8 +1,7 @@
 <form method="POST">
     {% csrf_token %}
-    Title:  {{ form.title }} <br>
-    Ingredients: {{ form.ingredients }} <button type="button" onclick="location.href='{% url 'ingredients:add' %}'" > Add ingredient</button> <br>
-    Instruction:  {{ form.instruction }} <br>
+  {{ form.as_p}}
   <input type="submit" value="Save">
   <input type="button" value="Reset" id="resetBtn" onClick="this.form.reset()" />
 </form>
+<a href="{% url 'recipies:createCategory' %}"><button>Create Category</button></a>
