+++ src/recipies/views.py
@@ -1,6 +1,6 @@
 from django.views import generic
 from django.urls import reverse_lazy
-from .models import Recipe
+from .models import Recipe, Category
 from django.shortcuts import get_object_or_404
 from django.http import HttpResponseRedirect
 from django.urls import reverse_lazy
@@ -16,7 +16,7 @@ class DetailView(generic.DetailView):
 
 class CreateView(generic.edit.CreateView):
     model = Recipe
-    fields = ["title", "ingredients", "instruction"]
+    fields = ["title", "ingredients", "instruction", "category"]
     success_url = reverse_lazy('recipies:list')
 
 
@@ -24,3 +24,9 @@ def delete(req, recipe_id):
     recipe = get_object_or_404(Recipe, pk=recipe_id)
     recipe.delete()
     return HttpResponseRedirect(reverse_lazy("recipies:list"))
+
+class CreateCategory(generic.CreateView):
+    model = Category
+    template_name = 'category_form.html'
+    fields = ["name"]
+    success_url = reverse_lazy("recipies:list")
